Use shared error component
